requirements.txt
torch==1.13.0
torchvision==0.14.0
numpy==1.23.5
opencv-python-headless==4.7.0.72
pillow==9.3.0
scipy==1.10.1
scikit-learn==1.3.2
tqdm==4.65.0
matplotlib==3.6.2
wandb==0.15.0
pyyaml==6.0
gitpython==3.1.4